Update repl API doc
